Updated PUT/DELETE request

+++ src/main/java/com/example/first/controller/BoardController.java
@@ -12,13 +12,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.http.MediaType;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -31,26 +30,18 @@ public class BoardController {
     private final UserRepository userRepository;
     // 게시글 등록
     @PostMapping("/players/{playerId}/boards")
-    public BoardDto create(
+    public ResponseEntity<String> create(
             Authentication auth,
             @PathVariable Long playerId,
-            @Valid @RequestBody BoardRequestDto request,
-            Authentication authentication
+            @Valid @RequestBody BoardRequestDto request
     ){
-     if (authentication == null || !authentication.isAuthenticated()) {
-
-     }
-     String currentUsername = authentication.getName();
-     User user = userRepository.findByUsername(currentUsername)
-             .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다"));
-     String displayAuthor = user.getEmail();
-     if (displayAuthor == null || displayAuthor.isEmpty()) {
-         displayAuthor = user.getUsername();
-     }
-     request.setAuthor(displayAuthor);
-     request.setPlayerId(playerId);
+        if (tryAuthAndSetName(auth, request)) return new ResponseEntity<>("인증되지 않은 사용자입니다.", HttpStatus.UNAUTHORIZED);
+        request.setPlayerId(playerId);
      BoardEntity saved = boardService.create(request);
-     return toDto(saved);
+     if(toDto(saved) != null) {
+         return new ResponseEntity<>("입력되었습니다.", HttpStatus.CREATED);
+     }
+     return new ResponseEntity<>("정보 처리 중 오류가 발생했습니다.", HttpStatus.BAD_REQUEST);
     }
 
     // 특정 선수 응원글 조회 추가
@@ -79,21 +70,44 @@ public class BoardController {
 
     // 게시글 수정
     @PutMapping("/{id}")
-    public BoardDto update(@PathVariable Long id, @RequestBody BoardRequestDto request) {
+    public ResponseEntity<String> update(Authentication auth,
+                                         @PathVariable Long id,
+                                         @RequestBody BoardRequestDto request) {
+        if (tryAuthAndSetName(auth, request)) return new ResponseEntity<>("인증되지 않은 사용자입니다.", HttpStatus.UNAUTHORIZED);
         BoardEntity updated = new BoardEntity(
                 request.getTitle(),
                 request.getContent(),
                 request.getAuthor(),
-                request.getMedia()
+                null
         );
-        return toDto(boardService.update(id, updated));
+        boardService.update(id, updated);
+        return new ResponseEntity<>("수정되었습니다.", HttpStatus.NO_CONTENT);
+        //return toDto(boardService.update(id, updated));
+    }
+
+    private boolean tryAuthAndSetName(Authentication auth, @RequestBody BoardRequestDto request) {
+        if (auth == null || !auth.isAuthenticated()) {
+            return true;
+        }
+        String currentUsername = auth.getName();
+        User user = userRepository.findByUsername(currentUsername)
+                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다"));
+        String displayAuthor = user.getEmail();
+        if (displayAuthor == null || displayAuthor.isEmpty()) {
+            displayAuthor = user.getUsername();
+        }
+        request.setAuthor(displayAuthor);
+        return false;
     }
 
     // 게시글 삭제
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable Long id) {
+    public ResponseEntity<String> delete(Authentication auth, @PathVariable Long id) {
+        if (auth == null || !auth.isAuthenticated()) {
+            return new ResponseEntity<>("인증되지 않은 사용자입니다.", HttpStatus.UNAUTHORIZED);
+        }
         boardService.delete(id);
-        return "삭제완료";
+        return new ResponseEntity<>("삭제되었습니다.", HttpStatus.NO_CONTENT);
     }
 
     // Entity -> Dto 변환
